requirements.txt
appnope==0.1.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.3
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
idna==2.8
imageio==2.6.1
importlib-metadata==1.5.2
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
Jinja2==2.11.1
json5==0.9.3
jsonschema==3.2.0
jupyter-client==6.1.1
jupyter-core==4.6.3
jupyterlab==2.0.1
jupyterlab-server==1.0.7
kaggle==1.5.6
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
notebook==6.0.3
numpy==1.17.3
pandas==0.25.2
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.1
prometheus-client==0.7.1
prompt-toolkit==3.0.4
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.2
pyrsistent==0.16.0
python-dateutil==2.8.0
python-slugify==4.0.0
pytz==2019.3
PyWavelets==1.1.1
pyzmq==19.0.0
requests==2.22.0
scikit-image==0.16.2
scipy==1.4.0
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
torch==1.4.0
torchvision==0.5.0
tornado==6.0.4
tqdm==4.41.0
traitlets==4.3.3
transforms==0.1
urllib3==1.24.3
wcwidth==0.1.9
webencodings==0.5.1
zipp==3.1.0
